Added priority decay scheduler, changed default to 5, max to 10

+++ proc.h
@@ -49,7 +49,7 @@ struct proc {
   struct file *ofile[NOFILE];  // Open files
   struct inode *cwd;           // Current directory
   char name[16];  // Process name (debugging)
-  float priority;  //New field
+  int priority;  //New field
 };
 
 // Process memory is laid out contiguously, low addresses first:
